% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_sage_client}
\alias{install_sage_client}
\title{Install the Python 'sage-data-client' Package}
\usage{
install_sage_client(envname = "r-sage-client")
}
\arguments{
\item{envname}{Name of the virtualenv to use or create (default: "r-sage-client")}
}
\description{
This function creates a virtualenv called 'r-sage-client' and installs the
Python dependency \code{sage-data-client} into it.
}
