% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sage_data.R
\name{query_sage_data}
\alias{query_sage_data}
\title{Query Sage Sensor Data}
\usage{
query_sage_data(start, end = NULL, filter = NULL, head = NULL, tail = NULL)
}
\arguments{
\item{start}{Required. Start timestamp (relative or absolute, in yyyy-mm-dd hh:mm:ss format).}

\item{end}{Optional. End timestamp.}

\item{filter}{A named list of filters (e.g., list(name = "env.temperature")).}

\item{head}{Optional. Limit to head values per series.}

\item{tail}{Optional. Limit to tail values per series.}
}
\value{
A data.frame containing the results.
}
\description{
This function wraps the Python function \code{sage_data_client.query()} to pull Sage/Crocus sensor data.
}
